--- src/GUI/RegisterForm.java
@@ -1,170 +1,0 @@
-package GUI;
-
-import javax.swing.*;
-import java.awt.*;
-import java.awt.event.*;
-
-public class RegisterForm extends JFrame {
-    private JTextField usernameField;
-    private JPasswordField passwordField;
-    private JPasswordField confirmPasswordField;
-    private JTextField emailField;
-    private JTextField fullNameField;
-    private JButton registerButton;
-    private JButton backButton;
-
-    public RegisterForm() {
-        setTitle("Movies Management System - Register");
-        setDefaultCloseOperation(JFrame.EXIT_ON_CLOSE);
-        setSize(500, 600);
-        setLocationRelativeTo(null);
-        setResizable(false);
-
-        // Main panel with gradient background
-        JPanel mainPanel = new JPanel() {
-            @Override
-            protected void paintComponent(Graphics g) {
-                super.paintComponent(g);
-                Graphics2D g2d = (Graphics2D) g;
-                g2d.setRenderingHint(RenderingHints.KEY_RENDERING, RenderingHints.VALUE_RENDER_QUALITY);
-                int w = getWidth();
-                int h = getHeight();
-                Color color1 = new Color(66, 139, 202);
-                Color color2 = new Color(51, 122, 183);
-                GradientPaint gp = new GradientPaint(0, 0, color1, w, h, color2);
-                g2d.setPaint(gp);
-                g2d.fillRect(0, 0, w, h);
-            }
-        };
-        mainPanel.setLayout(new BorderLayout());
-
-        // Title
-        JLabel titleLabel = new JLabel("Create New Account", SwingConstants.CENTER);
-        titleLabel.setFont(new Font("Arial", Font.BOLD, 24));
-        titleLabel.setForeground(Color.WHITE);
-        titleLabel.setBorder(BorderFactory.createEmptyBorder(30, 0, 20, 0));
-        mainPanel.add(titleLabel, BorderLayout.NORTH);
-
-        // Registration panel
-        JPanel registerPanel = new JPanel();
-        registerPanel.setOpaque(false);
-        registerPanel.setLayout(new GridBagLayout());
-        GridBagConstraints gbc = new GridBagConstraints();
-        gbc.gridwidth = GridBagConstraints.REMAINDER;
-        gbc.fill = GridBagConstraints.HORIZONTAL;
-        gbc.insets = new Insets(5, 20, 5, 20);
-
-        // Full Name field
-        addLabelAndField(registerPanel, "Full Name:", fullNameField = new JTextField(20), gbc);
-
-        // Username field
-        addLabelAndField(registerPanel, "Username:", usernameField = new JTextField(20), gbc);
-
-        // Email field
-        addLabelAndField(registerPanel, "Email:", emailField = new JTextField(20), gbc);
-
-        // Password field
-        addLabelAndField(registerPanel, "Password:", passwordField = new JPasswordField(20), gbc);
-
-        // Confirm Password field
-        addLabelAndField(registerPanel, "Confirm Password:", confirmPasswordField = new JPasswordField(20), gbc);
-
-        // Buttons panel
-        JPanel buttonPanel = new JPanel(new FlowLayout(FlowLayout.CENTER, 20, 0));
-        buttonPanel.setOpaque(false);
-
-        registerButton = new JButton("Register");
-        styleButton(registerButton);
-        buttonPanel.add(registerButton);
-
-        backButton = new JButton("Back to Login");
-        styleButton(backButton);
-        buttonPanel.add(backButton);
-
-        registerPanel.add(buttonPanel, gbc);
-
-        mainPanel.add(registerPanel, BorderLayout.CENTER);
-
-        // Add action listeners
-        registerButton.addActionListener(e -> handleRegistration());
-        backButton.addActionListener(e -> openLogin());
-
-        add(mainPanel);
-    }
-
-    private void addLabelAndField(JPanel panel, String labelText, JTextField field, GridBagConstraints gbc) {
-        JLabel label = new JLabel(labelText);
-        label.setForeground(Color.WHITE);
-        label.setFont(new Font("Arial", Font.PLAIN, 14));
-        panel.add(label, gbc);
-
-        styleTextField(field);
-        panel.add(field, gbc);
-    }
-
-    private void styleTextField(JTextField textField) {
-        textField.setPreferredSize(new Dimension(300, 35));
-        textField.setFont(new Font("Arial", Font.PLAIN, 14));
-        textField.setBorder(BorderFactory.createCompoundBorder(
-            BorderFactory.createLineBorder(new Color(200, 200, 200)),
-            BorderFactory.createEmptyBorder(5, 5, 5, 5)
-        ));
-    }
-
-    private void styleButton(JButton button) {
-        button.setPreferredSize(new Dimension(150, 40));
-        button.setFont(new Font("Arial", Font.BOLD, 14));
-        button.setBackground(new Color(51, 122, 183));
-        button.setForeground(Color.WHITE);
-        button.setFocusPainted(false);
-        button.setBorderPainted(false);
-        button.setCursor(new Cursor(Cursor.HAND_CURSOR));
-
-        button.addMouseListener(new MouseAdapter() {
-            public void mouseEntered(MouseEvent e) {
-                button.setBackground(new Color(40, 96, 144));
-            }
-            public void mouseExited(MouseEvent e) {
-                button.setBackground(new Color(51, 122, 183));
-            }
-        });
-    }
-
-    private void handleRegistration() {
-        String username = usernameField.getText();
-        String password = new String(passwordField.getPassword());
-        String confirmPassword = new String(confirmPasswordField.getPassword());
-        String email = emailField.getText();
-        String fullName = fullNameField.getText();
-
-        // Basic validation
-        if (username.isEmpty() || password.isEmpty() || email.isEmpty() || fullName.isEmpty()) {
-            JOptionPane.showMessageDialog(this, "Please fill in all fields!", 
-                "Registration Error", JOptionPane.ERROR_MESSAGE);
-            return;
-        }
-
-        if (!password.equals(confirmPassword)) {
-            JOptionPane.showMessageDialog(this, "Passwords do not match!", 
-                "Registration Error", JOptionPane.ERROR_MESSAGE);
-            return;
-        }
-
-        // TODO: Implement registration logic here
-        JOptionPane.showMessageDialog(this, "Registration successful! Please login.", 
-            "Success", JOptionPane.INFORMATION_MESSAGE);
-        openLogin();
-    }
-
-    private void openLogin() {
-        LoginForm loginForm = new LoginForm();
-        loginForm.setVisible(true);
-        this.dispose();
-    }
-
-    public static void main(String[] args) {
-        SwingUtilities.invokeLater(() -> {
-            new RegisterForm().setVisible(true);
-        });
-    }
-}
